+++ spark/src/main/scala/org/apache/spark/Plugins.scala
@@ -23,6 +23,7 @@ import java.{util => ju}
 import java.util.Collections
 
 import org.apache.spark.api.plugin.{DriverPlugin, ExecutorPlugin, PluginContext, SparkPlugin}
+import org.apache.spark.comet.ShimCometDriverPlugin
 import org.apache.spark.internal.Logging
 import org.apache.spark.internal.config.{EXECUTOR_MEMORY, EXECUTOR_MEMORY_OVERHEAD}
 
@@ -39,9 +40,7 @@ import org.apache.comet.{CometConf, CometSparkSessionExtensions}
  *
  * To enable this plugin, set the config "spark.plugins" to `org.apache.spark.CometPlugin`.
  */
-class CometDriverPlugin extends DriverPlugin with Logging {
-  import CometDriverPlugin._
-
+class CometDriverPlugin extends DriverPlugin with Logging with ShimCometDriverPlugin {
   override def init(sc: SparkContext, pluginContext: PluginContext): ju.Map[String, String] = {
     logInfo("CometDriverPlugin init")
 
@@ -51,12 +50,8 @@ class CometDriverPlugin extends DriverPlugin with Logging {
       } else {
         // By default, executorMemory * spark.executor.memoryOverheadFactor, with minimum of 384MB
         val executorMemory = sc.getConf.getSizeAsMb(EXECUTOR_MEMORY.key)
-        val memoryOverheadFactor =
-          sc.getConf.getDouble(
-            EXECUTOR_MEMORY_OVERHEAD_FACTOR,
-            EXECUTOR_MEMORY_OVERHEAD_FACTOR_DEFAULT)
-        val memoryOverheadMinMib =
-          sc.getConf.getLong(EXECUTOR_MIN_MEMORY_OVERHEAD, EXECUTOR_MIN_MEMORY_OVERHEAD_DEFAULT)
+        val memoryOverheadFactor = getMemoryOverheadFactor(sc.getConf)
+        val memoryOverheadMinMib = getMemoryOverheadMinMib(sc.getConf)
 
         Math.max((executorMemory * memoryOverheadFactor).toLong, memoryOverheadMinMib)
       }
@@ -97,15 +92,6 @@ class CometDriverPlugin extends DriverPlugin with Logging {
         conf.getBoolean(CometConf.COMET_EXEC_ENABLED.key, false)
     )
   }
-}
-
-object CometDriverPlugin {
-  // `org.apache.spark.internal.config.EXECUTOR_MEMORY_OVERHEAD_FACTOR` was added since Spark 3.3.0
-  val EXECUTOR_MEMORY_OVERHEAD_FACTOR = "spark.executor.memoryOverheadFactor"
-  val EXECUTOR_MEMORY_OVERHEAD_FACTOR_DEFAULT = 0.1
-  // `org.apache.spark.internal.config.EXECUTOR_MIN_MEMORY_OVERHEAD` was added since Spark 4.0.0
-  val EXECUTOR_MIN_MEMORY_OVERHEAD = "spark.executor.minMemoryOverhead"
-  val EXECUTOR_MIN_MEMORY_OVERHEAD_DEFAULT = 384L
 }
 
 /**
