+++ spark/src/main/spark-4.0/org/apache/spark/sql/comet/ShimDecimalPrecision.scala
@@ -1,0 +1,10 @@
+package org.apache.spark.sql.comet
+
+import org.apache.spark.sql.catalyst.expressions.Expression
+
+trait ShimDecimalPrecision {
+  object DecimalExpression {
+    def unapply(e: Expression): Option[(Int, Int)] =
+      org.apache.spark.sql.types.DecimalExpression.unapply(e)
+  }
+}
